pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = ""
authors = ["Anastasiia Olifir <a.o.olifir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"
ipython = "^8.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
